bot-app/runtime/requirements.txt
# ════════════════════════════════════════════════════════════════
# Microsoft 365 Agents SDK + supporting packages
# ════════════════════════════════════════════════════════════════

# M365 Agents SDK (Bot Framework successor)
microsoft-agents-activity==0.8.0
microsoft-agents-hosting-core==0.8.0
microsoft-agents-hosting-aiohttp==0.8.0
microsoft-agents-hosting-teams==0.8.0
microsoft-agents-authentication-msal==0.8.0

# Web framework (required by hosting-aiohttp)
aiohttp>=3.11.0

# Environment configuration
python-dotenv>=1.0.0

# System metrics for heartbeat
psutil>=5.9.0

# Microsoft Graph SDK (Team + channel creation)
msgraph-sdk>=1.0.0
azure-identity>=1.15.0

# Azure Storage (Queue + Blob) — Entra ID / RBAC auth
azure-storage-queue>=12.9.0
azure-storage-blob>=12.19.0

# Azure Resource Management (list resource groups for teardown selection)
azure-mgmt-resource>=23.0.0

# Model Context Protocol client for Microsoft Learn MCP queries
mcp>=1.12.4
httpx>=0.27.0
openai>=1.0.0
